+++ core/reader/LogFileReader.cpp
@@ -424,6 +424,7 @@ LogFileReader::LogFileReader(const string& projectName,
     mHostLogPathDir = hostLogPathDir;
     mHostLogPathFile = hostLogPathFile;
     mHostLogPath = PathJoin(hostLogPathDir, hostLogPathFile);
+    // mRealLogPath = mHostLogPath; fix it in 1.8
     mTailLimit = tailLimit;
     mLastFilePos = 0;
     mLastFileSize = 0;
@@ -469,6 +470,7 @@ LogFileReader::LogFileReader(const std::string& projectName,
     mHostLogPathDir = hostLogPathDir;
     mHostLogPathFile = hostLogPathFile;
     mHostLogPath = PathJoin(hostLogPathDir, hostLogPathFile);
+    // mRealLogPath = mHostLogPath; fix it in 1.8
     mTailLimit = tailLimit;
     mLastFilePos = 0;
     mLastFileSize = 0;
@@ -1119,8 +1121,7 @@ bool LogFileReader::UpdateFilePtr() {
                          ("open file succeeded, project", mProjectName)("logstore", mCategory)("config", mConfigName)(
                              "log reader queue name", mHostLogPath)("file device", ToString(mDevInode.dev))(
                              "file inode", ToString(mDevInode.inode))("file signature", mLastFileSignatureHash)(
-                             "real file path", mRealLogPath)("file size", mLastFileSize)("last file position",
-                                                                                         mLastFilePos));
+                             "real file path", mRealLogPath)("last file position", mLastFilePos));
                 return true;
             } else {
                 mLogFileOp.Close();
@@ -1151,9 +1152,9 @@ bool LogFileReader::UpdateFilePtr() {
             GloablFileDescriptorManager::GetInstance()->OnFileOpen(this);
             LOG_INFO(sLogger,
                      ("open file succeeded, project", mProjectName)("logstore", mCategory)("config", mConfigName)(
-                         "log reader queue name", mHostLogPath)("file device", ToString(mDevInode.dev))(
-                         "file inode", ToString(mDevInode.inode))("file signature", mLastFileSignatureHash)(
-                         "file size", mLastFileSize)("last file position", mLastFilePos));
+                         "log reader queue name",
+                         mHostLogPath)("file device", ToString(mDevInode.dev))("file inode", ToString(mDevInode.inode))(
+                         "file signature", mLastFileSignatureHash)("last file position", mLastFilePos));
             return true;
         } else {
             mLogFileOp.Close();
@@ -1282,6 +1283,11 @@ bool LogFileReader::CheckFileSignatureAndOffset(int64_t& fileSize) {
         }
     }
 
+    // If file size is 0 and filename is changed, we cannot judge if the inode is reused by signature,
+    // so we just recreate the reader to avoid filename mismatch
+    if (mLastFileSize == 0 && mRealLogPath != mHostLogPath) {
+        return false;
+    }
     fileSize = endSize;
     mLastFileSize = endSize;
     bool sigCheckRst = CheckAndUpdateSignature(string(firstLine), mLastFileSignatureHash, mLastFileSignatureSize);
@@ -2007,6 +2013,11 @@ LogFileReader::FileCompareResult LogFileReader::CompareToFile(const string& file
         sigStr[readSize] = '\0';
         uint64_t sigHash = mLastFileSignatureHash;
         uint32_t sigSize = mLastFileSignatureSize;
+        // If file size is 0 and filename is changed, we cannot judge if the inode is reused by signature,
+        // so we just recreate the reader to avoid filename mismatch
+        if (mLastFileSize == 0 && filePath != mHostLogPath) {
+            return FileCompareResult_SigChange;
+        }
         bool sigSameRst = CheckAndUpdateSignature(string(sigStr), sigHash, sigSize);
         if (!sigSameRst) {
             return FileCompareResult_SigChange;